Create collatz.js
This is a code which generates an image of the indian tricolour using collatz sequences. 
The Collatz Sequence is a simple mathematical sequence that starts with any positive integer. If the number is even, you divide it by 2. If it's odd, you multiply it by 3 and add 1. Then, you repeat this process with the new result. The conjecture suggests that no matter what positive integer you start with, you will eventually reach the number 1 through these steps.
I used a javascript framework called p5js for this project.